requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
ccxt==4.3.48
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pybit==5.7.0
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyTelegramBotAPI==4.19.2
python-dotenv==1.0.1
python-telegram-bot==21.3
redis==5.0.6
requests==2.32.3
sniffio==1.3.1
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
